[harshini]Add . Implementation of Algorithm Programs

+++ alg/MonthlyPay.cs
@@ -9,11 +9,11 @@ namespace AlgorithmPrograms
         public static void Month()
         {
             Console.WriteLine("enter loan amount");
-            double P = Utility.readInt();
+            double P = Utility.readDouble();
             Console.WriteLine("enter years");
-            double Y = Utility.readInt();
+            double Y = Utility.readDouble();
             Console.WriteLine("enter interest");
-            double R = Utility.readInt();
+            double R = Utility.readDouble();
             double r = R / (12 * 100);
             double n = 12 * Y;
             Console.WriteLine((P * R) / (1 - Math.Pow(1 + r, -n)));
